requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib
python-multipart
reportlab
bandit
python-dotenv